+++ plugin-compass/src/main/java/com/mapbox/maps/plugin/compass/CompassViewImpl.kt
@@ -81,7 +81,9 @@ open class CompassViewImpl
   override var isCompassEnabled: Boolean
     get() = isEnabled
     set(value) {
-      isEnabled = value
+      if (isEnabled != value) {
+        isEnabled = value
+      }
     }
 
   /**