+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Mapbox welcomes participation and contributions from everyone.
  
 ## Bug fixes 🐞
 * [compose] Queue viewport operations when the `MapViewportState` is no yet attached to the map, to avoid losing events.
-
+* Fix `LogoView.logoEnabled` not being in sync with `MapView.logo.enabled` state.
 
 # 11.4.0-rc.1 May 08, 2024
 ## Features ✨ and improvements 🏁